challenges/misc/Krusty/Bot/requirements.txt
discord~=2.3.2
python-dotenv~=1.0.0
setuptools~=65.5.0
pillow~=10.4.0
